// projects/dxp-p1-003/scripts/provision-schema/src/types/DataverseMetadata.ts
export interface ODataCollectionResponse<T> {
  readonly value: readonly T[];
  readonly '@odata.nextLink'?: string;
}

export interface PublisherRecord {
  readonly publisherid: string;
  readonly friendlyname: string;
  readonly customizationprefix: string;
}

export interface SolutionRecord {
  readonly solutionid: string;
  readonly uniquename: string;
  readonly version: string;
}